+++ src/components/Dashboard/Budgets/index.tsx
@@ -64,7 +64,7 @@ import { Box,
     const card = (
       <>
       <CardContent>
-        <Typography variant="h5" component="div">
+        <Typography variant="h5" component="div" sx={{ fontWeight: 'bold'}}>
           Budgets
         </Typography>
         <Divider />