+++ src/components/Dashboard/Goals/index.tsx
@@ -104,7 +104,7 @@ const Goals = () => {
     const card = (
       <>
       <CardContent>
-        <Typography variant="h5" component="div">
+        <Typography variant="h5" component="div" sx={{ fontWeight: 'bold'}}>
           Goals
         </Typography>
         <Divider />