+++ src/components/PlaidLink/index.tsx
@@ -10,6 +10,7 @@ import { ListItem,
     ListItemIcon,
     ListItemText,
     ListItemButton,
+    Button,
 } from '@mui/material';
 import { usePlaidLink } from 'react-plaid-link';
 
@@ -79,6 +80,74 @@ const PlaidLink = ({ user_id }) => {
             <ListItemText primary='Connect Bank' />
             </ListItemButton>
         </ListItem>
+    )
+}
+
+export const PlaidLinkInstitution = (props) => {
+    if (props.user_id === null || props.user_id === "Unauthorized") return <></>;
+    const [linkToken, setLinkToken] = useState(null);
+
+    const fetchToken = useCallback(async () => {
+        console.log('fetching plaid link token...');
+        try {
+            const config = {
+                method: "post",
+                headers: {
+                    'earmark-api-key': process.env.EARMARK_API_KEY,
+                },
+                params: {
+                    user_id: props.user_id
+                },
+                url:'/api/createLinkToken',
+            }
+            const response = await axios(config);
+            setLinkToken(response.data.linkToken);
+        } catch (error) {
+            console.log("FETCH LINK TOKEN FAILURE, inside PlaidLink", error);
+        }
+    }, []);
+
+    useEffect(() => {
+        fetchToken();
+    }, [fetchToken]);
+
+    const onSuccess = useCallback(async (publicToken, metadata) => {
+        console.log(metadata);
+        console.log('api key:', process.env.EARMARK_API_KEY);
+        console.log('inside onsuc cb')
+        const config = {
+            method: "post",
+            headers: {
+                'earmark-api-key': process.env.EARMARK_API_KEY,
+            },
+            params: {
+                user_id: props.user_id,
+                publicToken: publicToken,
+            },
+            url: '/api/exchangeLinkToken',
+        };
+        try {
+            const response = await axios(config);
+            console.log(response.data)
+        } catch (error) {
+            console.log("FETCH LINK TOKEN FAILURE, inside PlaidLink", error);
+        }
+    }, []);
+
+    const config = {
+        token: linkToken,
+        onSuccess,
+    }
+    const { open, exit, ready } = usePlaidLink(config);
+
+    return (
+        <Button
+          variant='contained'
+          onClick={() => open()}
+          disabled={!ready}
+        >
+        Connect Bank
+        </Button>
     )
 }